docker: move config RPCs to config.go
